console_chat/requirements.txt
aiofiles==0.4.0
python-dotenv==0.19.0
